move blockchain helpers to separate package
